Enhance Syncable and SyncMaster Interfaces to Support Customizable Attribute Synchronization
