+++ src/Contracts/BaseSyncable.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Stancl\Tenancy\Contracts;
+
+interface BaseSyncable
+{
+    public function getGlobalIdentifierKeyName(): string;
+
+    public function getGlobalIdentifierKey();
+
+    public function getCentralModelName(): string;
+
+    public function getSyncedAttributeNames(): array;
+
+    public function triggerSyncEvent();
+}
